Move CST lookup code to validation module
